Move more stuff into scheduling.py.
