+++ scheduling.py
@@ -1,4 +1,6 @@
 from enum import Enum
+from typing import Dict, List
+import nycdb.dataset
 
 
 class Schedule(Enum):
@@ -43,3 +45,39 @@ class Schedule(Enum):
         """
 
         return self.value
+
+
+# The names of all valid NYC-DB datasets.
+DATASET_NAMES: List[str] = list(nycdb.dataset.datasets().keys())
+
+# The default schedule for a dataset loader, if
+# otherwise unspecified.
+DEFAULT_SCHEDULE = Schedule.YEARLY
+
+DATASET_SCHEDULES: Dict[str, Schedule] = {
+    'oca': Schedule.DAILY,
+    'dobjobs': Schedule.DAILY,
+    'dob_complaints': Schedule.DAILY,
+    'dob_violations': Schedule.DAILY,
+    'ecb_violations': Schedule.DAILY,
+    'hpd_violations': Schedule.DAILY,
+    'oath_hearings': Schedule.DAILY,
+    'hpd_vacateorders': Schedule.EVERY_OTHER_DAY,
+    'hpd_registrations': Schedule.EVERY_OTHER_DAY,
+    'hpd_complaints': Schedule.EVERY_OTHER_DAY,
+    'dof_sales': Schedule.EVERY_OTHER_DAY,
+    'pad': Schedule.EVERY_OTHER_DAY,
+    'acris': Schedule.EVERY_OTHER_DAY
+}
+
+
+def get_schedule_for_dataset(dataset: str) -> Schedule:
+    return DATASET_SCHEDULES.get(dataset, DEFAULT_SCHEDULE)
+
+
+def sanity_check():
+    for dataset in DATASET_SCHEDULES:
+        assert dataset in DATASET_NAMES, f"'{dataset}' must be a valid NYCDB dataset name"
+
+
+sanity_check()
