+++ tests/test_aws_schedule_tasks.py
@@ -1,10 +1,6 @@
 import aws_schedule_tasks
 
 
-def test_sanity_check_works():
-    aws_schedule_tasks.sanity_check()
-
-
 def test_create_input_str_works():
     s = aws_schedule_tasks.create_input_str('foo', 'boop', True)
     assert 'foo' in s
